+++ services/traction/api/endpoints/models/tenant_workflow.py
@@ -3,7 +3,7 @@ from enum import Enum
 
 class TenantWorkflowTypeType(str, Enum):
     connection = "api.services.ConnectionWorkflow"
-    issuer = "api.services.IssuerWorkflow"
+    # issuer = "api.services.IssuerWorkflow"
     schema = "api.services.SchemaWorkflow"
     issue_cred = "api.services.IssueCredentialWorkflow"
     present_cred = "api.services.PresentCredentialWorkflow"
