+++ services/traction/api/services/__init__.py
@@ -1,3 +1,2 @@
-from .IssuerWorkflow import IssuerWorkflow  # noqa: F401
 from .IssueCredentialWorkflow import IssueCredentialWorkflow  # noqa: F401
 from .PresentCredentialWorkflow import PresentCredentialWorkflow  # noqa: F401
